Register-add hotel, add Kamar, view history
Register : error di bagian register jadi belum bisa navigate ke add hotel

add kamar : DONE

view history : tidak bisa membuat booking, jadi tidak bisa di test untuk view nya/ tidak muncul
