--- TambahHotel/TambahHotelPage.xaml
@@ -1,71 +1,0 @@
-﻿<basic:MyPage  xmlns:basic="clr-namespace:Velacro.UIElements.Basic;assembly=Velacro" x:Name="tambahHotelPage"         
-    x:Class="Hotelin_Desktop.TambahHotel.TambahHotelPage"
-      xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
-      xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-      xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006" 
-      xmlns:d="http://schemas.microsoft.com/expression/blend/2008" 
-      xmlns:local="clr-namespace:Hotelin_Desktop.TambahHotel"
-      mc:Ignorable="d" 
-      d:DesignHeight="450" d:DesignWidth="620"
-      Title="TambahHotelPage">
-
-    <Grid Background="White">
-        <Label 
-            Content="Tambah Hotel"
-            VerticalContentAlignment="Center"
-            HorizontalAlignment="Left" 
-            VerticalAlignment="Top"
-            Margin="20,20,0,0" 
-            Height="47" 
-            Width="242" 
-            FontSize="26" 
-            Foreground="#707070"
-            FontWeight="Bold"/>
-        <TextBox x:Name="nama_hotel_txt" 
-            HorizontalAlignment="Left" 
-            Height="30" Text="Nama Hotel" 
-            TextWrapping="Wrap" 
-            VerticalAlignment="Top" 
-            Width="285" 
-            Background="{x:Null}" 
-            BorderBrush="{x:Null}" 
-            FontSize="14" 
-            FontFamily="Montserrat" 
-            VerticalContentAlignment="Center" Margin="20,98,0,0"/>
-        <TextBox x:Name="lokasi_hotel_txt" 
-            HorizontalAlignment="Left" 
-            Height="30" Text="Lokasi Hotel" 
-            TextWrapping="Wrap" 
-            VerticalAlignment="Top" 
-            Width="285" 
-            Background="{x:Null}" 
-            BorderBrush="{x:Null}" 
-            FontSize="14" 
-            FontFamily="Montserrat" 
-            VerticalContentAlignment="Center" Margin="20,147,0,0"/>
-        <TextBox x:Name="deskripsi_hotel_txt" 
-            HorizontalAlignment="Left" 
-            Height="30" Text="Deskripsi Hotel" 
-            TextWrapping="Wrap" 
-            VerticalAlignment="Top" 
-            Width="285" 
-            Background="{x:Null}" 
-            BorderBrush="{x:Null}" 
-            FontSize="14" 
-            FontFamily="Montserrat" 
-            VerticalContentAlignment="Center" Margin="20,196,0,0"/>
-        <Button x:Name="tambah_hotel_btn" 
-            Content="TAMBAH HOTEL" 
-            HorizontalAlignment="Left" 
-            Height="38" 
-            Margin="20,245,0,0" 
-            VerticalAlignment="Top" 
-            Width="285" 
-            Background="#FF5399AF" 
-            FontSize="14" 
-            Foreground="White" 
-            BorderBrush="{x:Null}"/>
-    </Grid>
-    
-
-</basic:MyPage>
