+++ TambahHotel/TambahHotelWindow.xaml
@@ -1,0 +1,192 @@
+﻿<basic:MyWindow xmlns:basic="clr-namespace:Velacro.UIElements.Basic;assembly=Velacro"
+        x:Class="Hotelin_Desktop.TambahHotel.TambahHotelWindow"
+        xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+        xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+        xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
+        xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
+        xmlns:local="clr-namespace:Hotelin_Desktop.TambahHotel"
+        mc:Ignorable="d"
+        Title="TambahHotelWindow" Height="450" Width="800">
+    <Grid>
+        <Grid Background="White" Margin="20" >
+            <Grid.Effect>
+                <DropShadowEffect 
+                    BlurRadius="20" 
+                    Color="#FFDEDEDE" 
+                    ShadowDepth="1"/>
+            </Grid.Effect>
+
+            <TextBlock
+                x:Name="hotel_information_tb"
+                HorizontalAlignment="Left" 
+                Margin="248,21,0,0" 
+                TextWrapping="Wrap" 
+                Text="HOTEL INFORMATION" 
+                VerticalAlignment="Top" 
+                FontFamily="Montserrat" 
+                FontSize="22" 
+                FontWeight="Bold" 
+                Foreground="#FF707070"/>
+            <TextBox 
+                x:Name="nama_hotel_txt" 
+                HorizontalAlignment="Left" 
+                Height="34" 
+                Margin="42,71,0,0" 
+                TextWrapping="Wrap" 
+                Text="Nama Hotel" 
+                VerticalAlignment="Top" 
+                Width="321" 
+                RenderTransformOrigin="-0.132,0.621" 
+                VerticalContentAlignment="Center" 
+                Padding="3">
+                <TextBox.Resources>
+                    <Style TargetType="{x:Type Border}">
+                        <Setter Property="CornerRadius" Value="6"/>
+                    </Style>
+                </TextBox.Resources>
+            </TextBox>
+            <TextBox 
+                x:Name="lokasi_txt" 
+                HorizontalAlignment="Left" 
+                Height="34" 
+                Margin="42,115,0,0" 
+                TextWrapping="Wrap" 
+                Text="Lokasi" 
+                VerticalAlignment="Top" 
+                Width="321" 
+                RenderTransformOrigin="-0.132,0.621" 
+                VerticalContentAlignment="Center" 
+                Padding="3">
+                <TextBox.Resources>
+                    <Style TargetType="{x:Type Border}">
+                        <Setter Property="CornerRadius" Value="6"/>
+                    </Style>
+                </TextBox.Resources>
+            </TextBox>
+            <TextBox 
+                x:Name="fasilitas_txt" 
+                HorizontalAlignment="Left" 
+                Height="34" 
+                Margin="388,71,0,0" 
+                TextWrapping="Wrap" 
+                Text="Fasilitas" 
+                VerticalAlignment="Top" 
+                Width="321" 
+                RenderTransformOrigin="-0.132,0.621" 
+                VerticalContentAlignment="Center" 
+                Padding="3">
+                <TextBox.Resources>
+                    <Style TargetType="{x:Type Border}">
+                        <Setter Property="CornerRadius" Value="6"/>
+                    </Style>
+                </TextBox.Resources>
+            </TextBox>
+            <TextBox 
+                x:Name="harga_txt" 
+                HorizontalAlignment="Left" 
+                Height="34" 
+                Margin="42,159,0,0" 
+                TextWrapping="Wrap" 
+                Text="Harga" 
+                VerticalAlignment="Top" 
+                Width="321" 
+                RenderTransformOrigin="-0.132,0.621" 
+                VerticalContentAlignment="Center" 
+                Padding="3">
+                <TextBox.Resources>
+                    <Style TargetType="{x:Type Border}">
+                        <Setter Property="CornerRadius" Value="6"/>
+                    </Style>
+                </TextBox.Resources>
+            </TextBox>
+            <TextBox 
+                x:Name="rating_txt" 
+                HorizontalAlignment="Left" 
+                Height="34" 
+                Margin="388,115,0,0" 
+                TextWrapping="Wrap" 
+                Text="Rating" 
+                VerticalAlignment="Top" 
+                Width="321" 
+                RenderTransformOrigin="-0.132,0.621" 
+                VerticalContentAlignment="Center" 
+                Padding="3">
+                <TextBox.Resources>
+                    <Style TargetType="{x:Type Border}">
+                        <Setter Property="CornerRadius" Value="6"/>
+                    </Style>
+                </TextBox.Resources>
+            </TextBox>
+            <TextBox 
+                x:Name="deskripsi_txt"
+                HorizontalAlignment="Left" 
+                Height="79" 
+                Margin="42,206,0,0" 
+                TextWrapping="Wrap" 
+                Text="Deskripsi" 
+                VerticalAlignment="Top" 
+                Width="321" 
+                RenderTransformOrigin="-0.132,0.621" 
+                Padding="3,7,3,3">
+                <TextBox.Resources>
+                    <Style TargetType="{x:Type Border}">
+                        <Setter Property="CornerRadius" Value="6"/>
+                    </Style>
+                </TextBox.Resources>
+            </TextBox>
+            <Rectangle
+                AllowDrop="True"
+                Fill="White"
+                RadiusX="6"
+                RadiusY="6"
+                StrokeDashArray="8"
+                StrokeThickness="1"
+                StrokeDashCap="Round"
+                Stroke="#707070" 
+                Margin="388,159,233,89"/>
+            <StackPanel
+                Margin="388,159,233,89" 
+                HorizontalAlignment="Center"
+                VerticalAlignment="Center">
+                <Image 
+                    
+                    Stretch="Uniform"
+                    Height="28" Width="67"/>
+                <TextBlock
+                    Text="Pilih Gambar Hotel" 
+                    FontFamily="Montserrat" 
+                    VerticalAlignment="Center" 
+                    HorizontalAlignment="Center" 
+                    FontSize="7" 
+                    Margin="0,10,0,0"/>
+            </StackPanel>
+            <Label 
+                x:Name="file_name_lbl"
+                Content="File Name" 
+                HorizontalAlignment="Center" 
+                Margin="530,159,43,186" 
+                FontSize="12" 
+                Width="179"
+                VerticalAlignment="Center" 
+                VerticalContentAlignment="Center" Height="34"/>
+            <Button 
+                x:Name="register_btn" 
+                Content="REGISTER" 
+                HorizontalAlignment="Left" 
+                Margin="308,319,0,0" 
+                VerticalAlignment="Top" 
+                Width="166" 
+                Height="33" 
+                Foreground="White" 
+                BorderBrush="{x:Null}" 
+                Background="#FF5399AF">
+                <Button.Resources>
+                    <Style TargetType="{x:Type Border}">
+                        <Setter Property="CornerRadius" Value="6"/>
+                    </Style>
+                </Button.Resources>
+            </Button>
+        </Grid>
+
+    </Grid>
+</basic:MyWindow>
