+++ bin/Debug/userToken.txt
@@ -1,1 +1,1 @@
-eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9.eyJpc3MiOiJodHRwOlwvXC9sb2NhbGhvc3Q6ODAwMFwvYXBpXC91c2VyXC9sb2dpbiIsImlhdCI6MTYwODA0NzI4MiwiZXhwIjoxNjA4MTEyMDgyLCJuYmYiOjE2MDgwNDcyODIsImp0aSI6ImN1UnkybTdLTGZDSlRNZGMiLCJzdWIiOjMsInBydiI6IjIzYmQ1Yzg5NDlmNjAwYWRiMzllNzAxYzQwMDg3MmRiN2E1OTc2ZjcifQ.g5vdeijTjTnuOKPVKBNTUwhb9oxrpehseG07O031fWs
+eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9.eyJpc3MiOiJodHRwOlwvXC9sb2NhbGhvc3Q6ODAwMFwvYXBpXC91c2VyXC9sb2dpbiIsImlhdCI6MTYwODYxNjMxMCwiZXhwIjoxNjA4NjgxMTEwLCJuYmYiOjE2MDg2MTYzMTAsImp0aSI6ImhOVWxkMjlnRlBOT2NwMHMiLCJzdWIiOjMsInBydiI6IjIzYmQ1Yzg5NDlmNjAwYWRiMzllNzAxYzQwMDg3MmRiN2E1OTc2ZjcifQ.Iui3oWOZOoQX4Tu3-eAkKL5Z97l5GeRBHrOAYcRgxHg
